Improvements of next_start_code
Details
- ParseByteStream was specialized to allow the compiler to optimize
a large function, that contains a loop to the next_start_code.
