+++ vk_video_decoder/libs/NvVideoParser/include/VulkanVideoDecoder.h
@@ -140,7 +140,18 @@ public:
     virtual bool Deinitialize();
     virtual bool ParseByteStream(const VkParserBitstreamPacket *pck, size_t *pParsedBytes);
     template <SIMD_ISA T>
+
     bool ParseByteStreamSimd(const VkParserBitstreamPacket *pck, size_t *pParsedBytes);
+#if defined(__x86_64__) || defined (_M_X64)
+    bool ParseByteStreamAVX2(const VkParserBitstreamPacket* pck, size_t *pParsedBytes);
+    bool ParseByteStreamAVX512(const VkParserBitstreamPacket* pck, size_t *pParsedBytes);
+    bool ParseByteStreamSSSE3(const VkParserBitstreamPacket* pck, size_t *pParsedBytes);
+#elif defined(__aarch64__)
+    bool ParseByteStreamSVE(const VkParserBitstreamPacket* pck, size_t *pParsedBytes);
+    bool ParseByteStreamNEON(const VkParserBitstreamPacket* pck, size_t *pParsedBytes);
+#elif defined(__ARM_ARCH_7A__) || defined(_M_ARM64)
+    bool ParseByteStreamNEON(const VkParserBitstreamPacket* pck, size_t *pParsedBytes);
+#endif
     virtual bool GetDisplayMasteringInfo(VkParserDisplayMasteringInfo *) { return false; }
 
 protected:
