+++ vk_video_decoder/libs/NvVideoParser/src/NextStartCodeAVX512.cpp
@@ -4,8 +4,14 @@
 #include <cstdint>
 #include "VulkanVideoParserIf.h"
 #include "VulkanVideoDecoder.h"
+#include "ByteStreamParser.h"
 #include "NvVideoParser/nvVulkanVideoUtils.h"
 #include "NvVideoParser/nvVulkanVideoParser.h"
+
+bool VulkanVideoDecoder::ParseByteStreamAVX512(const VkParserBitstreamPacket* pck, size_t *pParsedBytes)
+{
+    return ParseByteStreamSimd<SIMD_ISA::AVX512>(pck, pParsedBytes);
+}
 
 template<>
 size_t VulkanVideoDecoder::next_start_code<SIMD_ISA::AVX512>(const uint8_t *pdatain, size_t datasize, bool& found_start_code)
